fix issues from PR; decouple LevelGenerator logic;
