--- app/src/pages/LevelGenerator/LevelGenerator.scss
@@ -1,13 +1,0 @@
-.generator-grid {
-    display: grid;
-    grid-template-columns: repeat(32, auto);
-    grid-template-rows: repeat(32, auto);
-}
-
-.generator-grid__cell {
-    width: 32px;
-    height: 32px;
-    box-sizing: border-box;
-    font-size: 1rem;
-    text-align: center;
-}
