+++ app/src/pages/index.ts
@@ -6,4 +6,3 @@ export * from './LeaderboardPage/LeaderboardPage';
 export * from './AuthorizationPage/AuthorizationPage';
 export * from './RegistrationPage/RegistrationPage';
 export * from './EditPage/EditPage';
-export * from './LevelGenerator/LevelGenerator';
