+++ app/src/services/LevelGenerator/LevelGenerator.ts
@@ -1,59 +1,16 @@
-import cloneDepp from 'lodash/cloneDeep';
+import cloneDeep from 'lodash/cloneDeep';
+
+import {allOneType, cellNeighborhoods, existAndEqual, randSort, sumOf} from '~/services/LevelGenerator/utils';
 
 import {Levels} from './__levels';
 
-import {BackgroundMap, BackgroundPart, BackgroundAsset, Level} from '../../store/Level/types';
 import {BaseType, GeneratorConfiguration, MapPartials} from './types';
+import {BackgroundMap, BackgroundPart, BackgroundAsset, Level} from '~/store/Level/types';
 
 const MAX_SUM = 9;
 
-const exist = (arr: number[][], i: number, j: number): boolean =>
-    i >= 0 && j >= 0 && arr[i] !== undefined && arr[i][j] !== undefined;
-
-const cellNeighborhoods = (i: number, j: number): number[][] => [
-    [i - 1, j - 1],
-    [i - 1, j],
-    [i - 1, j + 1],
-    [i, j - 1],
-    [i, j],
-    [i, j + 1],
-    [i + 1, j - 1],
-    [i + 1, j],
-    [i + 1, j + 1]
-];
-
-const existAndEqual = (arr: number[][], i: number, j: number, val: number): boolean =>
-    exist(arr, i, j) && arr[i][j] === val;
-
-const sumOf = (arr: number[][], i: number, j: number, val: number) => {
-    const data = cellNeighborhoods(i, j);
-
-    let sum = 0;
-    for (let k = 0; k < data.length; k += 1) {
-        const [x, y] = data[k];
-        if (exist(arr, x, y) && arr[x][y] === val) {
-            sum += 1;
-        }
-    }
-
-    return sum;
-};
-
-const allOneType = (arr: number[][], i: number, j: number, val: number): boolean => {
-    const data = cellNeighborhoods(i, j);
-
-    for (let k = 0; k < data.length; k += 1) {
-        const [x, y] = data[k];
-        if (exist(arr, x, y) && arr[x][y] !== val) {
-            return false;
-        }
-    }
-
-    return true;
-};
-
 function mapWall(inputLevel: number[][]): number[][] {
-    const level = cloneDepp(inputLevel);
+    const level = cloneDeep(inputLevel);
 
     for (let i = 0; i < level.length; i += 1) {
         for (let j = 0; j < level[i].length; j += 1) {
@@ -402,19 +359,6 @@ const mapToGameFormat = (level: number[][]): BackgroundMap => {
     }
 
     return res;
-};
-
-const randSort = (array: number[][]): number[][] => {
-    const arr = array;
-    for (let i = 0; i < arr.length - 1; i += 1) {
-        const rand = Math.floor(Math.random() * i);
-
-        const tmp = arr[i];
-        arr[i] = arr[rand];
-        arr[rand] = tmp;
-    }
-
-    return arr;
 };
 
 export class LevelGenerator {
