+++ app/src/store/Game/sagas.ts
@@ -2,6 +2,7 @@ import {ForkEffect, put, call, takeEvery, select} from 'redux-saga/effects';
 
 import {drawMap} from '~/core/bg.canvas';
 import {createGame, exitGame, loadLevel, pauseGame, play} from '~/core/engine';
+import {levelRandomGenerateAction} from '~/store/Level/actions';
 
 import {GameSetStateAction, GAME_ACTION_TYPES} from './types';
 import {AppStoreState} from '~/store/types';
@@ -19,6 +20,7 @@ function* gameSetStateWorker(action: GameSetStateAction) {
         }
         case 'INTERLUDE': {
             const nextAction = yield call(loadLevel);
+            yield put(levelRandomGenerateAction());
             yield put(nextAction);
             break;
         }
@@ -27,7 +29,6 @@ function* gameSetStateWorker(action: GameSetStateAction) {
             const {map} = state.level.levels[state.game.level.number ?? 0];
 
             yield call(drawMap.bind(null, map));
-
             yield call(play);
             break;
         }
